requirements.txt
fitz==0.0.1.dev2
groq==0.8.0
langchain==0.2.3
langchain_community==0.2.4
langchain_core==0.2.5
langchain_groq==0.1.5
llama_index==0.10.43
pdf2image==1.17.0
Pillow==10.3.0
pytesseract==0.3.10
python-dotenv==1.0.1
streamlit==1.24.0
